frontend-bot/requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.5.0
sentencepiece
cmake